backend/app/requirements.txt
fastapi
uvicorn
httpx
pydantic
email-validator
python-dotenv